Fixed selection glitch under Qt
When clicking on the account selection arrow in an account sheet, the
entry table would always end up losing the focus after a short delay.
That was because the automatic cell editing feature in the sheet would
kick in.

I've fixed this by disabling the call to the super mouse press handler
in the case where we hit a custom decoration.