(#402 #14 #15) Drag & Drop / Amount values custom painting bug
This commit fixes the issue with painting amount values by _not_ calling
the base class paint method in instances where a value painter is present.
This removes the need to suppress the table model from hiding that display
information.  The usage of customized ItemFlags which facilitated that
suppression of the model was removed.  Fixes #402 issue with drag and drop
because now the model provides the DisplayRole data again.