Decouple click position from selection in Qt decoration click handlers
Previously, we would always, in the core, perform action triggered by
decoration clicks (show account and toggle excluded) on selected items.

There would be a problem when clicking a decoration in an unselected
row, but we would work around the problem by updating selection right
before clicking would occur.

It was dirty and fragile, and it broke in the 2.8.0 release.

In this commit, we uncouple selection from clicks. Click position is
always passed as an argument when a decoration is clicked and relevant
core methods have been enhanced to optionnally take that click position
as an argument (we still use selection when that argument isn't
supplied).

Fixes #404