Fix Qt warnings
We would get multiple "QTimer can only be used with threads started with
QThread" when running moneyGuru. Our sheet font size would also be a bit
glitchy.

I'm not entirely sure why, but it was fixed by breaking the huge
circular references going on with the app singleton and all its
sheets/tables. Now, the global reference allowing sheets/tables to be
notified when prefs are updated is the `Preferences` instance itself.

We still use a global variable because it would be really (and
needlessly, IMHO) complicated to get this reference down to every table
in the app.