Fix import regression from v2.9
The first problem in the regression is obvious: a dangling mis-reference
(`probability` instead of `weight`).

There is, however, a second underlying problem: That part of the code is
not supposed to be invoked without non-core import binding plugins
because the reference binder (which only binds through txn reference
number) does not do any fuzzy matching.

We would mistakenly end up executing that part of the code because we
would treat `None` as a valid match element, either from the "existing"
or "import" side.

So we fix both these problems at once here, with tests.

fixes #432