fix #442: Amounts of different currencies can't be compared
Analysis: This error comes from loading older (or hand-edited) files that contain reconciled splits whose amount currency doesn't match their account's currency.  The current versions of MoneyGuru do a good job of preventing users from reconciling such off-currency splits, and also prevents changing an account's currency if doing so would produce off-currency reconciliations, so we believe this problem presents only in files created with older versions.

Fix: When loading saved data, we look for off-currency splits and ignore any saved reconciliation state.  While this produces an in-memory image that differs from the saved state, we don't mark the loaded file as dirty, because the dirty state implies that closing without save and then re-opening the original file would produce a different state, and it won't in this case.  The on-disk reconciliation data is considered to be invalid, and (as of this fix) can never be loaded.

Additional:
1. Any splits that do not have an account but are marked reconciled are also caught by this fix.  We have no evidence that there are files with such splits; there certainly shouldn't be!
2. The currency of splits with a 0 amount is irrelevant, and a reconciled 0 split with an explicit currency that doesn't match the account remains reconciled when it is loaded.
3. There is a new test and a new test data file for this fix.