Support Python 3.5 on OS X
We do this by running WAF with python2.7 instead of with the same executable as the one being used with moneyGuru.

We had to de-frameworkize psmtabbarcontrol because its script imported code from hscommon which is in python3. A drawback of this is that we now have all images stuck in the same place. An upside is that we get rid of some complexity.