Refuse to load moneyguru files containing unsupported currencies
With the new plugin management feature coming, it becomes easy to
mistakenly disable currency providers. If we do that and then load a
file containing a now-unsupported currency, we would previously silently
use our default currency. That's bad because the user could then save
the file and lose all currency information. We really don't want this to
happen.

ref #451