qt: wait until the main window is initialized before opening a document
Previously, we would open the document before our main window was
initialized. When the opening would result in an error, that would leave
the app in a buggy state.

We now load the initial document after the initialization of the main
window and properly create a new document in cases where the opening of
the initial document failed.