qt: get rid of QSignalMapper in base panel
Under some versions of PyQt5, the `QWidget` overload of the `mapped`
signal of our `QSignalMapper` would refuse to work. I couldn't easily
find a way to make it work.

Because the use of a signal mapper wasn't such a great idea in the first
place, I zapped it and replaced it with good old individual signal maps
for each widget in the panel.

fixes #456